requirements.txt
discord.py
urllib3
requests
